[Fix 🪛] input type속성 추가에 따른 타입 오류 해결
