--- src/pages/Home/constants/Constants.ts
@@ -1,10 +1,0 @@
-export const SEARCH_INPUTS = {
-  search: {
-    validate: {
-      required: { value: true, message: "" },
-    },
-    attributes: {
-      placeholder: "어떤 운동이 좋을까요?",
-    },
-  },
-}
