--- src/pages/Login/constants/Constants.ts
@@ -1,22 +1,0 @@
-export const LOGIN_INPUTS = {
-  DEFAULT_VALUES: {
-    userName: "",
-    password: "",
-  },
-  userName: {
-    attributes: {
-      placeholder: "아이디",
-    },
-    validate: {
-      required: { value: true },
-    },
-  },
-  password: {
-    attributes: {
-      placeholder: "비밀번호",
-    },
-    validate: {
-      required: { value: true },
-    },
-  },
-}
